temp sensor fix

+++ components/MTS4x/MTS4x.cpp
@@ -79,7 +79,7 @@ float MTS4X::readTemperature(bool waitOnNewVal) {
   uint8_t lsb = Wire.read();
   uint8_t msb = Wire.read();
 
-  uint16_t rawTemp = ((uint16_t)msb << 8) | lsb;
+  int16_t rawTemp = (msb << 8) | lsb;
 
   return MTS4X_RAW_TO_CELSIUS(rawTemp);
 }
